requirements.txt
dj-database-url==0.4.1
Django==1.9.8
gunicorn==19.6.0
psycopg2==2.6.2
whitenoise==3.2
spotipy==2.3.8
Pillow==3.3.0
django-environ==0.41